requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.10.0
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg==1.20.0
ecdsa==0.17.0
feedgenerator==1.9.2
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Material==0.1.1
ghp-import==2.0.1
idna==3.2
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mdx-video==0.1.8
node==0.9.26
numpy==1.21.1
odict==1.7.0
packaging==21.0
pandas==1.3.1
paramiko==2.7.2
pelican==3.6.0
pelican-alias==1.1
Pillow==8.3.1
plumber==1.6
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
pyflakes==2.3.1
Pygments==2.10.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scipy==1.7.1
seaborn==0.11.1
Serializer==0.2.1
simplejson==3.17.5
six==1.16.0
smartypants==2.0.1
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
text-unidecode==1.3
typogrify==2.0.7
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
